refactor(server): extract cluster_queries.go
Move cluster/group CRUD and auto-detection upsert functions.

Part of #75.
